refactor(domain): Distance 원시값 포장
